code refactor
* moved demangled_string with const char* parameter to any.cpp
* moved helper functions to type_internal scope
