+++ src/shogun/lib/any.h
@@ -44,12 +44,12 @@
 #include <string>
 #include <typeinfo>
 #include <type_traits>
-#ifdef HAVE_CXA_DEMANGLE
-#include <cxxabi.h>
-#endif
 
 namespace shogun
 {
+	std::string demangled_type_helper(const char* name);
+	std::string demangled_type(const char* name);
+
 	/** Converts compiler-dependent name of class to
 	 * something human readable.
 	 * @return human readable name of class
@@ -57,17 +57,9 @@ namespace shogun
 	template <typename T>
 	std::string demangled_type()
 	{
-#ifdef HAVE_CXA_DEMANGLE
-		size_t length;
-		int status;
-		char* demangled =
-		    abi::__cxa_demangle(typeid(T).name(), nullptr, &length, &status);
-		std::string demangled_string(demangled);
-		free(demangled);
-#else
-		std::string demangled_string(typeid(T).name());
-#endif
-		return demangled_string;
+		const char* name = typeid(T).name();
+
+		return demangled_type_helper(name);
 	}
 
 	enum class PolicyType
